Refactor fs into submodules
